wip feat: script to automate releases
The script retrieve the next version from tools/next-version, then
populate the annotations to include a changelog based on the list of
commits, and create a tag based with the generated annotations.

The intention is remove manual intervention when generating a release,
leting the developer to focus more on the changes to do and organize the
changes.

This auto-release expect the repository follow convention for
conventional commits.

For every type of commit we have to duplicate th grep expression to
admit the broken changes.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>